- changed return value in four classes of *DiscountService

+++ src/main/java/core/basesyntax/strategy/impl/BlackFridayDiscountService.java
@@ -5,6 +5,6 @@ import core.basesyntax.strategy.DiscountService;
 public class BlackFridayDiscountService implements DiscountService {
     @Override
     public double getDiscount() {
-        return (double) 45;
+        return 45;
     }
 }
